Update Main
Easing into the Pokemon battle side of things, and also fixing a bug I ran into. Basically, a new Starter variable was being created in the local frame and all the necessary values were being added to it. This made my script end prematurely as my code was trying to read the Starter variable in the global frame. I solved this issue by creating a global variable in the PokePick function with the same name, solving the creation of the new variable. Then I also started to work on the hp and abilities of the Starters and also a basic script which prints out the list as well.